services/altdata-source/requirements.txt
# 追加刚才安装的 clickhouse-connect
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.4
pydantic-settings==2.2.1
httpx==0.27.0
clickhouse-connect==0.7.6
pyyaml==6.0.1
beautifulsoup4==4.12.3
